Some cleanup...

+++ MODSRC/vazkii/botania/client/render/tile/RenderTileSpreader.java
@@ -55,10 +55,10 @@ public class RenderTileSpreader extends TileEntitySpecialRenderer {
 
 		Minecraft.getMinecraft().renderEngine.bindTexture(spreader.isRedstone() ? textureRs : spreader.isDreamwood() ? textureDw : texture);
 		GL11.glScalef(1F, -1F, -1F);
-		
+
 		if(spreader.isULTRA_SPREADER()) {
-			Color color = Color.getHSBColor((float) ((((double) System.currentTimeMillis() / 10) + new Random(spreader.xCoord ^ spreader.yCoord ^ spreader.zCoord).nextInt(10000)) % 360) / 360F, 0.4F, 0.9F);
-			GL11.glColor3f((float) color.getRed() / 255F, (float) color.getGreen() / 255F, (float) color.getBlue() / 255F);
+			Color color = Color.getHSBColor((float) (((double) System.currentTimeMillis() / 10 + new Random(spreader.xCoord ^ spreader.yCoord ^ spreader.zCoord).nextInt(10000)) % 360) / 360F, 0.4F, 0.9F);
+			GL11.glColor3f(color.getRed() / 255F, color.getGreen() / 255F, color.getBlue() / 255F);
 		}
 		model.render();
 		GL11.glColor3f(1F, 1F, 1F);
